refactor(#218): refactor Canvas Page
