+++ apps/frontend/src/features/drawing/hooks/index.ts
@@ -9,3 +9,6 @@ export * from './useConstraints';
 export * from './useViewport';
 export * from './useSchemaEditor';
 export * from './useMemoStore';
+export * from './useCanvasInitialization';
+export * from './useCanvasKeyboard';
+export * from './useCanvasNodes';
